+++ argtyped/__init__.py
@@ -1,3 +1,3 @@
-from .__version__ import VERSION as __version__
-from .arguments import *
-from .custom_types import Choices, Enum, Switch, auto
+from argtyped.__version__ import VERSION as __version__
+from argtyped.arguments import *
+from argtyped.custom_types import Enum, Switch, auto