+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0]
+### Changed
+- Removed `Choices` in favor of `Literal`.
+- Changed packaging and tool configurations to using `pyproject.toml`.
 
 ## [0.3.1]
 ### Added